Folder structure is changed and router is added
