+++ index.ts
@@ -1,6 +1,7 @@
 import express, { Express, Response, Request } from "express";
-import TheatreWithMovieControllerInstance from "./components/BMS/controllers/theatre.controller";
-import HeartbeatControllerInstance from "./components/heartbeat/Controllers/heartbeat.Controller";
+import HeartbeatControllerInstance from "./src/controllers/heartbeat.controller";
+import TheatreRouter from "./src/routes/theatre.route";
+
 const PORT = 3000;
 const app: Express = express();
 
@@ -9,10 +10,7 @@ app.get("/", (req: Request, res: Response) => {
 });
 
 app.get("/heartbeat", HeartbeatControllerInstance.getTimeStamp);
-app.get(
-  "/getTheatreWithShowTime",
-  TheatreWithMovieControllerInstance.getTheatreAndShowTimeWithMovie
-);
+app.use("/getTheatreWithShowTime", TheatreRouter);
 
 app.listen(process.env.NODE_ENV || PORT, () => {
   // eslint-disable-next-line no-console
