+++ src/models/show.model.ts
