+++ src/routes/theatre.route.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import theatreWithMovieController from "../controllers/theatre.controller";
+
+const router = Router();
+
+router
+  .route("/")
+  .get([theatreWithMovieController.getTheatreAndShowTimeWithMovie]);
+
+export default router;
